make an explicit erlmode derived from erlang-mode
Its always a bad idea to copy an existing mode not rename it. The
right way to do this is to derive a new mode from the existing mode
and the change the *new* mode. This patch setes the stage for that. It
creates the new 'erlmode' derived from erlang mode and points the
various start hooks to that. Figuring out what changes exist from
erlang mode and porting them to the new derived mode is left as a
future excersize.